Allow script to be started from any location
The script was failing when it was started from Task Scheduler because
it was setup with relative paths. The use of Resolve-Path finds the
absolute path based on the the current working directory, not the
directory the script is in. Therefore, it was not successful in building
absolute paths based on the relative paths in the script. However,
the $PSScriptRoot variable contains the absolute path of the script,
regardless of where it was invoked.